Demo of Chris Force Directed Graph
This demo is a force directed graph that draws data from the Chris API. Currently the weights of the graph are skewed due to the amount of data being accessed causing the edges to be extremely long and causing the data to be clustered. Our next steps are fix the weigh problem and to implement more UI friendly nodes. These new nodes will give the user quick access information, display any images contained in the node, and express the state of each node. The Chris UI uses various states such as: waiting, starting, computing, receiving, registering, and finished.